changes woked fine, put a handler into remove dir to only remove if exists

+++ attachments.py
@@ -25,4 +25,5 @@ def Generate(fExt,minfSize,maxfSize,noOfAttachments):
     GenerateAttachments(fExt,minfSize,maxfSize,noOfAttachments)
 
 def Remove():
-    shutil.rmtree('temp')
+    if os.path.isdir('temp'):
+        shutil.rmtree('temp')
